had to make os specific changes

+++ linux/kern_dynamic.h
@@ -9,7 +9,7 @@ public:
     // virtual methods to be implemented, 
     int parseAndGetGlobals();
     void insert_section(std::string sec_name, uint64_t sh_offset, uint64_t sh_size);
-    int ksym_dlsym(const char* newString, size_t* out_address);
+    int ksym_dlsym(const char* newString, uint64_t* out_address);
     
     int kcrc_index(std::string symbol, uint32_t* kcrc);
