+++ windows/kern_static.h
@@ -15,8 +15,8 @@ public:
     void insert_section(std::string sec_name, uint64_t sh_offset, uint64_t sh_size) {};
     int insert_sections();
 
-    std::map<std::string, size_t>* kern_sym_map_fetch() { return &kern_sym_map; };
-    std::map<std::string, size_t>* kern_off_map_fetch()  { return &kern_off_map; };
+    std::map<std::string, uint64_t>* kern_sym_map_fetch() { return &kern_sym_map; };
+    std::map<std::string, uint64_t>* kern_off_map_fetch()  { return &kern_off_map; };
 
     kernel_windows* get_syskern_static() { return (kernel_windows*)this; };
 private:
