+++ xnu/kern_dynamic.h
@@ -8,7 +8,7 @@ class kern_dynamic : public kernel_xnu
 {
 public:
     // virtual methods to be implemented, 
-    int ksym_dlsym(const char* newString, size_t* out_address);
+    int ksym_dlsym(const char* newString, uint64_t* out_address);
     int parseAndGetGlobals();
     void insert_section(std::string sec_name, uint64_t sh_offset, uint64_t sh_size);
