+++ xnu/kern_static.h
@@ -8,13 +8,13 @@ class kern_static : public kernel_xnu
 {
 public:
     // virtual methods to be implemented, 
-    int ksym_dlsym(const char* newString, size_t* out_address);
+    int ksym_dlsym(const char* newString, uint64_t* out_address);
     int parseAndGetGlobals();
     void insert_section(std::string sec_name, uint64_t sh_offset, uint64_t sh_size) {};
     int insert_sections();
 
-    std::map<std::string, size_t>* kern_sym_map_fetch() { return &kern_sym_map; };
-    std::map<std::string, size_t>* kern_off_map_fetch()  { return &kern_off_map; };
+    std::map<std::string, uint64_t>* kern_sym_map_fetch() { return &kern_sym_map; };
+    std::map<std::string, uint64_t>* kern_off_map_fetch()  { return &kern_off_map; };
 
 private:
     using kernel_xnu::kernel_xnu;
